Update Aéroport image and fix agency ID in ligne.json
Replaced references to img/A.png with img/Aeroport.png in agences.html and added the new image file. Deleted img/1.png and updated img/A.png. Modified index.html to handle a special case for the 'GC' agency image path. Changed the agency ID from 1 to 'A' in ligne.json for consistency.